Clear completed task function added
